outro/requirements.txt
absl-py==2.0.0
altair==5.1.2
anyio==4.0.0
astor==0.8.1
astunparse==1.6.3
attrdict==2.0.1
attrs==23.1.0
Babel==2.13.0
bce-python-sdk==0.8.92
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.4
decorator==5.1.1
easyocr==1.7.1
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
future==0.18.3
gast==0.5.4
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.6
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lmdb==1.4.1
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat
paddleocr
paddlepaddle
pandas==2.1.1
pdf2docx==0.5.6
Pillow==10.0.1
premailer==3.10.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycryptodome==3.19.0
pydeck==0.8.1b0
Pygments==2.16.1
PyMuPDF==1.20.2
pyparsing==3.1.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
rarfile==4.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ultralytics==8.0.200
unicode==2.9
Unidecode==1.3.7
urllib3==2.0.7
validators==0.22.0
visualdl==2.5.3
watchdog==3.0.0
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0